Remove circular build dependency. Ensures the library form of the website never depends on generated content.
