+++ gridsome.server.js
@@ -6,7 +6,6 @@
 // To restart press CTRL + C in terminal and run `gridsome develop`
 const express = require('express');
 const path = require('path');
-const merge = require('webpack-merge')
 
 const $AssetsJSON = require("./data/assets.json");
 
@@ -20,7 +19,7 @@ module.exports = function (api) {
   });
 
   // for gridsome develop
-  api.configureServer((app) => {
+  api.configureServer(app => {
     // webpack won't host these files in development mode.
     app.use('/admin/lib', express.static(path.join(__dirname, 'admin/lib')));
