--- src/pages/projects/Orientation.vue
@@ -1,9 +1,0 @@
-<template>
-  <dti-project project="orientation" />
-</template>
-
-<script lang="ts">
-import Vue from 'vue';
-
-export default Vue.extend({});
-</script>
