--- src/pages/projects/Shout.vue
@@ -1,9 +1,0 @@
-<template>
-  <dti-project project="shout" />
-</template>
-
-<script lang="ts">
-import Vue from 'vue';
-
-export default Vue.extend({});
-</script>
