+++ src/shims-image-loader.d.ts
@@ -1,3 +1,5 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
+
 declare module '*.svg' {
   const any: any;
   export default any;
