+++ src/shims-style-loader.ts
@@ -1,3 +1,5 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
+
 declare module '*.css' {
   const any: any;
   export default any;
