+++ src/views/projects/QueueMeIn.vue
@@ -1,0 +1,16 @@
+<template>
+  <dti-project project="queuemein" />
+</template>
+
+<script lang="ts">
+import Vue from 'vue';
+import { Component, Prop } from 'vue-property-decorator';
+
+import Strings from '@/strings/strings';
+
+@Component({})
+export default class QueueMeIn extends Vue {
+  @Prop({ required: true })
+  Strings!: Strings;
+}
+</script>
